Create Caching

+++ Entities/Concrete/Car.cs
@@ -14,7 +14,7 @@ namespace Entities.Concrete
         public int ColordId { get; set; }
         public string Name { get; set; }
         public string ModelYear { get; set; }
-        public decimal DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public string Description { get; set; }
